+++ CHANGES.md
@@ -23,6 +23,8 @@
 - Add ForbiddenInheritanceCheck.
 - TrailingWhitespaceCheck now marks the actual whitespace.
 - Add defintion functionality to magik-language-server.
+- Move functionality from implementation provider to definition provider in magik-language-server.
+- Implementation provider now provides impementations of abstract methods.
 - Various small fixes.
 
 0.7.1 (2023-02-21)